+++ src/main/java/io/github/torand/fastersql/order/Order.java
@@ -21,5 +21,7 @@ import io.github.torand.fastersql.Sql;
 import java.util.stream.Stream;
 
 public interface Order extends Sql {
+    String alias();
+
     Stream<Field> fieldRefs();
 }