+++ CHANGELOG.md
@@ -10,8 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Statement parameters of type OffsetDateTime and ZonedDateTime are converted into Timestamp.
-- supporting the 'current_date' and 'current_time' system functions.
-
+- Supporting the 'current_date' and 'current_time' system functions.
+- Supporting the 'round' scalar function
+- Validates that order aliases are specified in the SELECT clause.
+ 
 ### Changed
 - ...